recommendations/requirements.txt
grpcio-tools~= 1.30
grpc-interceptor
pytest
